Move all files into src
