+++ package.json
@@ -8,14 +8,14 @@
     "lint": "eslint .",
     "build": "tsc --esModuleInterop"
   },
+  "dependencies": {
+    "node-fetch": "^2.6.1"
+  },
   "devDependencies": {
     "@paralenz/eslint-config-typescript-react": "^1.0.3",
     "@types/node": "^16.11.7",
-    "@types/node-fetch": "^3.0.3",
+    "@types/node-fetch": "^2.5.10",
     "typescript": "^4.4.4"
-  },
-  "dependencies": {
-    "node-fetch": "^2.6.1"
   },
   "keywords": [
     "autocomplete",
