+++ src/constants.ts
@@ -1,0 +1,7 @@
+import { GeoSearchConfig } from '.'
+
+export const DEFAULT_CONFIG: GeoSearchConfig = {
+  language: 'en'
+}
+
+export const OK = 200
